greybus: gb-audio: Add integrated greybus audio driver
So this is still an early attempt at merging my dummy driver
into the greybus driver code. Things are coming along, but
its not ready for merging yet.

This patch requires the kernel config include the
SND_GB_AUDIO patch I also submitted since it requires
the simple card infrastructure and codec support.

Since folks may not test this on a tegra, I added spdif
codec support, which is baiscally an empty codec. We can
switch back to the rt56xx codecs when we start talking
to actual modules.

I've gotten things tied in with the greybus initialization
logic, and somewhat tested via gbsim, although there are
still bugs with module removal.

The patch still doesn't handle audio transmission properly,
currently using a fixed rate timer to send chunks of audio.
This will need to be tweaked so we're sending the right
chunks of audio at the right interval for the given bitrate.

This patch was generated on top of the queue MarkG sent
me, and while I don't think it depends on anything he
added, it may not apply cleanly w/o.

Also credit to Mark Greer, who has contributed a number of
changes to this that have been folded in.

Signed-off-by: John Stultz <john.stultz@linaro.org>